Login: refactoring
