+++ packages/login/src/app/(root)/page.tsx
@@ -24,65 +24,28 @@
 // content are licensed under the terms of the Creative Commons Attribution-ShareAlike 4.0
 // International. See the License terms at http://creativecommons.org/licenses/by-sa/4.0/legalcode
 
-"use server";
-
-import { redirect } from "next/navigation";
-import { cookies } from "next/headers";
-
-import { getBgPattern } from "@docspace/shared/utils/common";
-import { SYSTEM_THEME_KEY } from "@docspace/shared/constants";
-import { ThemeKeys } from "@docspace/shared/enums";
-
 import Login from "@/components/Login";
-import { LoginFormWrapper } from "@/components/Login/Login.styled";
-import {
-  getSettings,
-  getThirdPartyProviders,
-  getCapabilities,
-  getSSO,
-  getColorTheme,
-} from "@/utils/actions";
+import { getSettings } from "@/utils/actions";
+import LoginForm from "@/components/LoginForm";
+import ThirdParty from "@/components/ThirdParty";
+import RecoverAccess from "@/components/RecoverAccess";
 
-async function Page({
-  searchParams,
-}: {
-  searchParams: { [key: string]: string };
-}) {
-  const [settings, thirdParty, capabilities, ssoSettings, colorTheme] =
-    await Promise.all([
-      getSettings(),
-      getThirdPartyProviders(),
-      getCapabilities(),
-      getSSO(),
-      getColorTheme(),
-    ]);
-
-  const ssoUrl = capabilities ? capabilities.ssoUrl : "";
-  const hideAuthPage = ssoSettings ? ssoSettings.hideAuthPage : false;
-
-  if (ssoUrl && hideAuthPage && searchParams.skipssoredirect !== "true") {
-    redirect(ssoUrl);
-  }
-
-  const bgPattern = getBgPattern(colorTheme?.selected);
-
-  const cookieStore = cookies();
-
-  const systemTheme = cookieStore.get(SYSTEM_THEME_KEY);
+async function Page() {
+  const settings = await getSettings();
 
   return (
-    <LoginFormWrapper id="login-page" bgPattern={bgPattern}>
-      <div className="bg-cover" />
-      <Login
-        searchParams={searchParams}
-        capabilities={capabilities}
-        settings={typeof settings === "string" ? undefined : settings}
-        thirdPartyProvider={thirdParty}
-        ssoSettings={ssoSettings}
-        isAuthenticated={false}
-        systemTheme={systemTheme?.value as ThemeKeys}
-      />
-    </LoginFormWrapper>
+    <Login>
+      {settings && typeof settings !== "string" && (
+        <>
+          <LoginForm
+            hashSettings={settings?.passwordHash}
+            cookieSettingsEnabled={settings?.cookieSettingsEnabled}
+          />
+          <ThirdParty />
+          {settings.enableAdmMess && <RecoverAccess />}
+        </>
+      )}
+    </Login>
   );
 }
