+++ packages/login/src/components/Login/Login.context.tsx
@@ -24,9 +24,24 @@
 // content are licensed under the terms of the Creative Commons Attribution-ShareAlike 4.0
 // International. See the License terms at http://creativecommons.org/licenses/by-sa/4.0/legalcode
 
-import { useEffect, useLayoutEffect } from "react";
+"use client";
 
-const canUseDOM = typeof window !== "undefined";
-const useIsomorphicLayoutEffect = canUseDOM ? useLayoutEffect : useEffect;
+import React, { createContext, useState } from "react";
 
-export default useIsomorphicLayoutEffect;
+export const LoginLoadingValueContext = createContext(false);
+
+export const LoginLoadingDispatchContext = createContext(
+  (value: boolean) => {},
+);
+
+export const LoginContext = ({ children }: { children: React.ReactNode }) => {
+  const [isLoading, setIsLoading] = useState(false);
+
+  return (
+    <LoginLoadingDispatchContext.Provider value={setIsLoading}>
+      <LoginLoadingValueContext.Provider value={isLoading}>
+        {children}
+      </LoginLoadingValueContext.Provider>
+    </LoginLoadingDispatchContext.Provider>
+  );
+};
