+++ packages/login/src/hooks/useTheme.ts
@@ -1,3 +1,4 @@
+import { i18n } from "i18next";
 // (c) Copyright Ascensio System SIA 2009-2024
 //
 // This program is a free software product.
@@ -30,21 +31,17 @@ import { Base, Dark, TColorScheme, TTheme } from "@docspace/shared/themes";
 import { getEditorTheme, getSystemTheme } from "@docspace/shared/utils";
 import { ThemeKeys } from "@docspace/shared/enums";
 import { getAppearanceTheme } from "@docspace/shared/api/settings";
-import { TGetColorTheme, TSettings } from "@docspace/shared/api/settings/types";
+import { TGetColorTheme } from "@docspace/shared/api/settings/types";
 import { setCookie } from "@docspace/shared/utils/cookie";
 import { SYSTEM_THEME_KEY } from "@docspace/shared/constants";
 
-import useI18N from "./useI18N";
-
 export interface UseThemeProps {
   colorTheme?: TGetColorTheme;
-  settings?: TSettings;
   systemTheme?: ThemeKeys;
+  i18n: i18n;
 }
 
-const useTheme = ({ colorTheme, settings, systemTheme }: UseThemeProps) => {
-  const { i18n } = useI18N({ settings });
-
+const useTheme = ({ colorTheme, systemTheme, i18n }: UseThemeProps) => {
   const [currentColorTheme, setCurrentColorTheme] =
     React.useState<TColorScheme>({} as TColorScheme);
