+++ packages/login/src/utils/registry.tsx
@@ -53,4 +53,3 @@ export default function StyledComponentsRegistry({
     </StyleSheetManager>
   );
 }
-
