+++ packages/shared/components/scrollbar/Scrollbar.tsx
@@ -24,6 +24,8 @@
 // content are licensed under the terms of the Creative Commons Attribution-ShareAlike 4.0
 // International. See the License terms at http://creativecommons.org/licenses/by-sa/4.0/legalcode
 
+"use client";
+
 import React, { useEffect, useRef, useState } from "react";
 import { useTheme } from "styled-components";
